+++ src/main/java/net/frozenblock/wilderwild/tag/WilderBiomeTags.java
@@ -63,6 +63,8 @@ public final class WilderBiomeTags {
 	public static final TagKey<Biome> HAS_FALLEN_PALM_AND_JUNGLE_AND_OAK = bind("has_fallen_palm_and_jungle_and_oak");
 	public static final TagKey<Biome> HAS_FALLEN_LARGE_JUNGLE = bind("has_fallen_large_jungle");
 	public static final TagKey<Biome> HAS_COMMON_FALLEN_LARGE_JUNGLE = bind("has_common_fallen_large_jungle");
+	public static final TagKey<Biome> HAS_FALLEN_DARK_OAK = bind("has_fallen_dark_oak");
+	public static final TagKey<Biome> HAS_COMMON_FALLEN_DARK_OAK = bind("has_common_fallen_dark_oak");
 	public static final TagKey<Biome> HAS_FALLEN_BIRCH_AND_OAK_DARK_FOREST = bind("has_fallen_birch_and_oak_dark_forest");
 	public static final TagKey<Biome> HAS_FALLEN_SPRUCE_TREES = bind("has_fallen_spruce_trees");
 	public static final TagKey<Biome> HAS_CLEAN_FALLEN_SPRUCE_TREES = bind("has_clean_fallen_spruce_trees");